+++ src/tools/layer2/index.ts
@@ -5,14 +5,19 @@ import { registerBindToChannel } from "./bindToChannel.js";
 import { registerConnectNodes } from "./connectNodes.js";
 import { registerCreateContainer } from "./createContainer.js";
 import { registerCreateControlPanel } from "./createControlPanel.js";
+import { registerCreateControlSurface } from "./createControlSurface.js";
 import { registerCreateExternalIo } from "./createExternalIo.js";
 import { registerCreateGlslShader } from "./createGlslShader.js";
+import { registerCreateMacro } from "./createMacro.js";
 import { registerCreateNodeChain } from "./createNodeChain.js";
+import { registerCreatePhoneRemote } from "./createPhoneRemote.js";
 import { registerCreatePythonScript } from "./createPythonScript.js";
 import { registerDuplicateNetwork } from "./duplicateNetwork.js";
 import { registerManageCheckpoint } from "./manageCheckpoint.js";
 import { registerManageComponent } from "./manageComponent.js";
+import { registerManageCue } from "./manageCue.js";
 import { registerManagePresets } from "./managePresets.js";
+import { registerRandomizeControls } from "./randomizeControls.js";
 import { registerSetParametersBatch } from "./setParametersBatch.js";
 
 export const layer2Registrars: ToolRegistrar[] = [
@@ -23,11 +28,16 @@ export const layer2Registrars: ToolRegistrar[] = [
   registerSetParametersBatch,
   registerCreateContainer,
   registerCreateControlPanel,
+  registerCreateControlSurface,
   registerAnimateParameter,
   registerBindToChannel,
   registerManagePresets,
   registerManageCheckpoint,
+  registerManageCue,
   registerManageComponent,
+  registerCreateMacro,
+  registerRandomizeControls,
+  registerCreatePhoneRemote,
   registerCreateExternalIo,
   registerDuplicateNetwork,
   registerArrangeNetwork,